Remove package-lock.json and test-auth middleware; refactor imports in various components to use absolute paths. Clean up firebase-auth middleware by removing development logging. Update userService import statement for consistency.
