+++ backend/src/middleware/firebase-auth.js
@@ -7,9 +7,6 @@ export const verifyToken = async (req, res, next) => {
     }
     try {
         const decodedToken = await admin.auth().verifyIdToken(token);
-        if(process.env.NODE_ENV === 'development') {
-            console.log("Decoded token:", {uid: decodedToken.uid});
-        }
         req.user = decodedToken;
         next();
     } catch (error) {
