+++ frontend/app/components/home/HomePage.tsx
@@ -1,9 +1,9 @@
 import { BarChart, LineChart, PieChart } from "lucide-react"
 import { Link } from "react-router"
 
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "~/components/ui/card"
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "~/components/ui/tabs"
-import { Button } from "~/components/ui/button"
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card"
+import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
+import { Button } from "@/components/ui/button"
 
 export function HomePage() {
     return (
