+++ frontend/app/components/recipes/RecipeList.tsx
@@ -5,9 +5,9 @@ import { Button } from "@/components/ui/button"
 import { Input } from "@/components/ui/input"
 import { ScrollArea } from "@/components/ui/scroll-area"
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
-import { DialogContent, DialogTrigger } from "../ui/dialog"
-import { Dialog } from "../ui/dialog"
-import CreateRecipe from "./createRecipe"
+import { DialogContent, DialogTrigger } from "@/components/ui/dialog"
+import { Dialog } from "@/components/ui/dialog"
+import CreateRecipe from "@/components/recipes/createRecipe"
 
 // Sample recipe data
 const recipes = [
