+++ frontend/app/pages/auth/register.tsx
@@ -1,4 +1,4 @@
-import { RegisterForm } from "../../components/auth/RegisterForm"
+import { RegisterForm } from "@/components/auth/RegisterForm"
 export default function RegisterPage() {
     return (
         <div className="flex min-h-svh w-full items-center justify-center p-6 md:p-10">
