+++ frontend/app/pages/dashboard/page.tsx
@@ -1,4 +1,4 @@
-import Dashboard from "~/components/dashboard/Dashboard";
+import Dashboard from "@/components/home/HomePage";
 
 function DashboardPage() {
     return (
