+++ frontend/app/service/authService.ts
@@ -1,6 +1,6 @@
 import { fetchSignInMethodsForEmail, createUserWithEmailAndPassword, signInWithEmailAndPassword, signOut, type User, onAuthStateChanged, updateProfile, updateEmail, updatePassword } from "firebase/auth";
-import auth from "~/config/firebaseConfig";
-import type { UpdateUser } from "~/model/user";
+import auth from "@/config/firebaseConfig";
+import type { UpdateUser } from "@/model/user";
 
 export interface AuthCredentials {
     email: string;
@@ -113,4 +113,4 @@ export const authService = {
     }
 
 }
-
+export default authService;
