+++ frontend/app/service/recipeSerive.ts
@@ -1,6 +1,6 @@
-import type { Recipe } from "~/model/recipe";
-import { authService } from "./authService";
-const recipeService = {
+import type { Recipe } from "@/model/recipe";
+import authService from "@/service/authService";
+export const recipeService = {
     createRecipe: async (recipe: Recipe) => {
         try {
             const token = await authService.getJwtToken();
